Begin building out AIFN library
