+++ mkdocs.yml
@@ -22,8 +22,10 @@ nav:
       - Loaders: guide/concepts/loaders_and_documents.md
       - Infrastructure: guide/concepts/infra.md
       - Plugins: guide/concepts/plugins.md
+    - AI Functions:
+      - Data: guide/ai_functions/data.md
+      - Strings: guide/ai_functions/strings.md
     - Use Cases:
-      - Data cleaning: guide/use_cases/data_cleaning.md
       - Enforcing LLM output formats: guide/use_cases/enforcing_format.md
       - Slackbot: guide/use_cases/slackbot.md
     - Development:
