+++ src/marvin/ai_functions/__init__.py
@@ -1,1 +1,3 @@
 from .base import ai_fn
+
+from . import data, strings
