+++ src/marvin/ai_functions/strings.py
@@ -1,0 +1,8 @@
+from marvin.ai_functions import ai_fn
+
+
+@ai_fn
+def title_case(input: str) -> str:
+    """
+    Given a string {input} change the case to make it APA style guide title case.
+    """
